Split metastore unit tests into files
