+++ pramen-py/src/pramen_py/models/__init__.py
@@ -41,6 +41,12 @@ class TableFormat(enum.Enum):
     delta: str = "delta"
 
 
+@enum.unique
+class Query(enum.Enum):
+    path: str = enum.auto()
+    table: str = enum.auto()
+
+
 @attrs.define(auto_attribs=True, frozen=True, slots=True)
 class InfoDateSettings:
     column: str = attrs.field()
